Moves artificial data routines to `artificial.py`
- additionally fixes test_feature.py `compare`, so that it closely
resembles the original
- some PEP8 things
