+++ trackpy/tests/test_feature.py
@@ -14,165 +14,26 @@ from numpy.testing import assert_almost_equal, assert_allclose
 from numpy.testing.decorators import slow
 from pandas.util.testing import (assert_series_equal, assert_frame_equal,
                                  assert_produces_warning)
-from scipy.spatial import cKDTree
 
 import trackpy as tp
 from trackpy.try_numba import NUMBA_AVAILABLE
-from trackpy.utils import validate_tuple
+from trackpy.artificial import (draw_feature, draw_spots, draw_point,
+                                gen_nonoverlapping_locations)
 
 
 path, _ = os.path.split(os.path.abspath(__file__))
-
-
-def feat_gauss(r, rg=0.333):
-    """ Gaussian at r = 0 with max value of 1. Its radius of gyration is
-    given by rg. """
-    return np.exp((r/rg)**2 * r.ndim/-2)
-
-
-def feat_gauss_edge(r, value_at_edge=0.1):
-    """ Gaussian at r = 0 with max value of 1. Its value at r = 1 is given by
-    value_at_edge. """
-    return np.exp(np.log(value_at_edge)*r**2)
-
-
-def feat_ring(r, r_at_max, value_at_edge=0.1):
-    """ Ring feature with a gaussian profile, centered at r_at_max. Its value 
-    at r = 1 is given by value_at_edge."""
-    return np.exp(np.log(value_at_edge)*((r - r_at_max) / (1 - r_at_max))**2)
-
-
-def feat_hat(r, disc_size, value_at_edge=0.1):
-    """ Solid disc of size disc_size, with Gaussian smoothed borders. """
-    mask = r > disc_size
-    spot = (~mask).astype(r.dtype)
-    spot[mask] = feat_ring(r[mask], disc_size, value_at_edge)
-    spot[~mask] = 1
-    return spot
-
-
-def feat_step(r):
-    """ Solid disc. """
-    return r <= 1
-
-
-def draw_feature(image, position, diameter, max_value=None,
-                 feat_func=feat_gauss, ecc=None, **kwargs):
-    """ Draws a radial symmetric feature and adds it to the image at given
-    position.
-
-    Parameters
-    ----------
-    image : ndarray
-        image to draw features on
-    position : iterable
-        coordinates of feature position
-    diameter : number
-        defines the box that will be drawn on
-    max_value : number
-        maximum feature value. should be much less than the max value of the
-        image dtype, to avoid pixel wrapping at overlapping features
-    feat_func : function. Default: feat_gauss
-        function f(r) that takes an ndarray of radius values
-        and returns intensity values <= 1
-    ecc : positive number, optional
-        eccentricity of feature, defined only in 2D. Identical to setting
-        diameter to (diameter / (1 - ecc), diameter * (1 - ecc))
-    kwargs : keyword arguments are passed to feat_func
-    """
-    if len(position) != image.ndim:
-        raise ValueError("Number of position coordinates should match image"
-                         " dimensionality.")
-    diameter = validate_tuple(diameter, image.ndim)
-    if ecc is not None:
-        if len(diameter) != 2:
-            raise ValueError("Eccentricity is only defined in 2 dimensions")
-        if diameter[0] != diameter[1]:
-            raise ValueError("Diameter is already anisotropic; eccentricity is"
-                             " not defined.")
-        diameter = (diameter[0] / (1 - ecc), diameter[1] * (1 - ecc))
-    radius = tuple([(d - 1) / 2 for d in diameter])
-    if max_value is None:
-        max_value = np.iinfo(image.dtype).max - 3
-    rect = []
-    vectors = []
-    for (c, r, lim) in zip(position, radius, image.shape):
-        if (c >= lim) or (c < 0):
-            raise ValueError("Position outside of image.")
-        lower_bound = max(int(np.floor(c - r)), 0)
-        upper_bound = min(int(np.ceil(c + r + 1)), lim)
-        rect.append(slice(lower_bound, upper_bound))
-        vectors.append(np.arange(lower_bound - c, upper_bound - c) / r)
-    coords = np.meshgrid(*vectors, indexing='ij', sparse=True)
-    r = np.sqrt(np.sum(np.array(coords)**2, axis=0))
-    spot = max_value * feat_func(r, **kwargs)
-    image[rect] += spot.astype(image.dtype)
-
-
-def draw_point(image, pos, value):
-    image[tuple(pos)] = value
-
-
-def gen_random_locations(shape, count, margin=0):
-    margin = validate_tuple(margin, len(shape))
-    np.random.seed(0)
-    pos = [np.random.randint(round(m), round(s - m), count)
-           for (s, m) in zip(shape, margin)]
-    return np.array(pos).T
-
-
-def eliminate_overlapping_locations(f, separation):
-    separation = validate_tuple(separation, f.shape[1])
-    assert np.greater(separation, 0).all()
-    # Rescale positions, so that pairs are identified below a distance of 1.
-    f = f / separation
-    while True:
-        duplicates = cKDTree(f, 30).query_pairs(1)
-        if len(duplicates) == 0:
-            break
-        to_drop = []
-        for pair in duplicates:
-            to_drop.append(pair[1])
-        f = np.delete(f, to_drop, 0)
-    return f * separation
-
-
-def gen_nonoverlapping_locations(shape, count, separation, margin=0):
-    positions = gen_random_locations(shape, count, margin)
-    return eliminate_overlapping_locations(positions, separation)
-
-
-def draw_spots(shape, positions, diameter, noise_level=0, bitdepth=8,
-               feat_func=feat_gauss, **kwargs):
-    if bitdepth <= 8:
-        dtype = np.uint8
-        internaldtype = np.uint16
-    elif bitdepth <= 16:
-        dtype = np.uint16
-        internaldtype = np.uint32
-    elif bitdepth <= 32:
-        dtype = np.uint32
-        internaldtype = np.uint64
-    else:
-        raise ValueError('Bitdepth should be <= 32')
-    np.random.seed(0)
-    image = np.random.randint(0, noise_level + 1, shape).astype(internaldtype)
-    for pos in positions:
-        draw_feature(image, pos, diameter, max_value=2**bitdepth - 1,
-                     feat_func=feat_func, **kwargs)
-    return image.clip(0, 2**bitdepth - 1).astype(dtype)
 
 
 def compare(shape, count, radius, noise_level, engine):
     radius = tp.utils.validate_tuple(radius, len(shape))
     # tp.locate ignores a margin of size radius, take 1 px more to be safe
     margin = tuple([r + 1 for r in radius])
-    diameter = tuple([(r + 1) * 2 for r in radius])
+    diameter = tuple([(r * 2) + 1 for r in radius])
+    draw_range = tuple([d * 3 for d in diameter])
     cols = ['x', 'y', 'z'][:len(shape)][::-1]
-    pos = gen_nonoverlapping_locations(shape, count, diameter, margin)
-    image = draw_spots(shape, pos, diameter, noise_level)
-    f = tp.locate(image, [2*r + 1 for r in radius],
-                  engine=engine)
+    pos = gen_nonoverlapping_locations(shape, count, draw_range, margin)
+    image = draw_spots(shape, pos, draw_range, noise_level)
+    f = tp.locate(image, diameter, engine=engine)
     actual = f[cols].sort(cols)
     expected = DataFrame(pos, columns=cols).sort(cols)
     return actual, expected
@@ -373,7 +234,7 @@ class CommonFeatureIdentificationTests(object):
         actual = locate(np.flipud(image.transpose()))
         assert len(actual)
 
-    def test_subpx_precision(self): 
+    def test_subpx_precision(self):
         self.check_skip()
         L = 21
         dims = (L, L + 2)  # avoid square images in tests
@@ -413,7 +274,7 @@ class CommonFeatureIdentificationTests(object):
         expected = DataFrame(np.asarray(pos).reshape(1, -1), columns=cols)
         assert_allclose(actual, expected, atol=PRECISION)
 
-        pos = [7.75, 13]  # center is between pixels, biased right 
+        pos = [7.75, 13]  # center is between pixels, biased right
         image = np.ones(dims, dtype='uint8')
         draw_point(image, pos1, 50)
         draw_point(image, pos2, 100)
@@ -433,7 +294,7 @@ class CommonFeatureIdentificationTests(object):
         expected = DataFrame(np.asarray(pos).reshape(1, -1), columns=cols)
         assert_allclose(actual, expected, atol=PRECISION)
 
-        pos = [7, 12.75]  # center is between pixels, biased up 
+        pos = [7, 12.75]  # center is between pixels, biased up
         image = np.ones(dims, dtype='uint8')
         draw_point(image, pos1, 50)
         draw_point(image, pos2, 100)
